requirements.txt
discord.py
markovify
spacy
spacytextblob
psutil
better_profanity
python-dotenv
dotenv
pillow
watchdog
py-cpuinfo
websocket-client
unalix==0.9
psutil
requests
easyocr
pytubefix
python-slugify
requests-async